backend/scrapers/requirements.txt
beautifulsoup4==4.11.2
lxml[html_clean]
pandas==2.2.2
python-dotenv==1.0.1
requests_html==0.10.0
supabase==2.6.0
